useUser methods, nuxt-authorization
