+++ app/plugins/authorization-resolver.ts
@@ -7,8 +7,7 @@ export default defineNuxtPlugin({
         authorization: {
           resolveClientUser: async () => {
             // Your logic to retrieve the user from the client
-            const { data: session } = await authClient.useSession(useFetch)
-            return session?.value?.user
+            return (await useUser())?.value
           },
         },
       },
