+++ app/utils/authClient.ts
@@ -1,6 +1,6 @@
 import { inferAdditionalFields } from "better-auth/client/plugins"
 import { createAuthClient } from "better-auth/vue"
-import type { auth } from "~~/server/utils/auth"
+import { toRef } from "vue"
 
 /**
  * authClient
@@ -8,3 +8,12 @@ import type { auth } from "~~/server/utils/auth"
 export const authClient = createAuthClient({
   plugins: [inferAdditionalFields<typeof auth>()],
 })
+
+/**
+ * getUser
+ */
+export async function useUser() {
+  const { data: session } = await authClient.useSession(useFetch)
+  const user = toRef(() => sessionToUserDetail(session?.value))
+  return user
+}
