+++ server/plugins/authorization-resolver.ts
@@ -1,11 +1,10 @@
+import { useUser } from "../utils/auth"
+
 export default defineNitroPlugin((nitroApp) => {
   nitroApp.hooks.hook("request", async (event) => {
     event.context.$authorization = {
       resolveServerUser: async () => {
-        const session = await auth.api.getSession({
-          headers: event.headers,
-        })
-        return session?.user
+        return useUser(event)
       },
     }
   })
